# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Lobster",
    "description": "Go to /login/ to obtain your token. Authorize with the following value: 'Token {your_token}'"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": [

      ]
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "paths": {
...
    "/courses/<courseid>/admins/": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets list of admins for a course",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Add admins to a course",
        "requestBody": {
          "description": "List of admins emails",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Remove admins from a course",
        "requestBody": {
          "description": "List of admins emails",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/courses/<courseid>/staff/": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets list of staff for a course",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Add staff to a course",
        "requestBody": {
          "description": "List of staff emails",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Remove staff from a course",
        "requestBody": {
          "description": "List of staff emails",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/courses/<courseid>/students/": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets list of students for a course",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Add students to a course",
        "requestBody": {
          "description": "List of student emails",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "courseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Remove students from a course",
        "requestBody": {
          "description": "List of student emails",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "users": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/<userid>/courses_is_admin_for/": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userid",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Gets list of courses the user is an admin for",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/<userid>/courses_is_staff_for/": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userid",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Gets list of courses the user is an admin for",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/<userid>/courses_is_student_for/": {
      "get": {
        "tags": [
          "Permissions"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userid",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Gets list of courses the user is an admin for",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/users/<userid>/": {
      "get": {
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userid",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Gets information on a specific user",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}